Update linux.rst
Elasticsearch install instructions. Changed paragraph about Ubuntu not
providing a package and that the most recent, stable version should be used.
This is contrary to what the IMPORTANT note below it states, about AtoM only
working with 0.90.11. I fell foul of this by jumping ahead and
downloading/installing the latest version off the Elasticsearch web site, only
to read further on that the newer version isn't supported. My wording could
perhaps be improved, but you get the idea.