Remove reference to 'suggested packages'
'Suggested' packages associated with a given package are not by default installed alongside that package. The previous wording kind of implied that they would be, and that, to prevent this behaviour, --no-install-recommends is required. The --install-suggests option is generally needed to install suggested packages. This commit removes the references to suggested packages.