Refactor asynch jobs page, fixes #188
In the updated 2.7 installation documentation, we have moved the majority
of the installation and configuration steps for the job scheduler from its
current separate page right into the main Linux Ubutu installation
documentation. However with this change, the existing "Asynchronous
jobs and worker management" documentation page in the Admin manual was made
redundant. This commit will refactor the previous page, moving it to the
"Maintenance" section of the Admin manual, and changing its focus to
managing jobs and the atom-worker.